Misra C adherence

+++ fw/EF_GPIO8.h
@@ -35,13 +35,13 @@
 /******************************************************************************
 * Macros and Constants
 ******************************************************************************/
-#define GPIO8_INPUT 0
-#define GPIO8_OUTPUT 1
+#define GPIO8_INPUT     ((uint32_t)0)
+#define GPIO8_OUTPUT    ((uint32_t)1)
 
-#define EF_GPIO8_DATAI_MAX_VALUE    0x000000FF         // Maximum value of the DATAI register, 8 bits
-#define EF_GPIO8_DATAO_MAX_VALUE    0x000000FF         // Maximum value of the DATAO register, 8 bits
-#define EF_GPIO8_DIR_MAX_VALUE      0x000000FF         // Maximum value of the DIR register, 8 bits
-#define EF_GPIO8_NUM_PINS           0x00000008         // Number of GPIO pins
+#define EF_GPIO8_DATAI_MAX_VALUE    ((uint32_t)0x000000FF)         // Maximum value of the DATAI register, 8 bits
+#define EF_GPIO8_DATAO_MAX_VALUE    ((uint32_t)0x000000FF)         // Maximum value of the DATAO register, 8 bits
+#define EF_GPIO8_DIR_MAX_VALUE      ((uint32_t)0x000000FF)         // Maximum value of the DIR register, 8 bits
+#define EF_GPIO8_NUM_PINS           ((uint32_t)0x00000008)         // Number of GPIO pins
 
 /******************************************************************************
 * Typedefs and Enums
